Add prev/next page, add/remove transaction, totals
Prev and Next buttons will cause only the appropriate Transactions to render.
If you click a Transaction, it will be deleted. For now, you click anywhere on the Transaction to remove it; later we will add an X or trash icon.
You can also add a Transaction to the list by pressing the Add Transaction button. It will create a Transaction with some random values.
Finally, a running total of all the positive and all the negative Transactions will be shown up top. This will give the users a quick idea on how much they've spent compared to how much money they've put into their budget tracking.